no-config osf-metrics
